Move shaders to new file
