+++ src/server/network/server.cpp
@@ -16,13 +16,13 @@ Server::Server()
     // Create spawn area
     for (int cz = 0; cz < TEMP_WORLD_SIZE; cz++) {
         for (int cx = 0; cx < TEMP_WORLD_SIZE; cx++) {
-                std::array<int, CHUNK_AREA> heightMap = makeHeightMap({cx, 0, cz});
+            std::array<int, CHUNK_AREA> heightMap = makeHeightMap({cx, 0, cz});
 
-                for (int cy = 0; cy < 5; cy++) {
-                    Chunk &chunk = m_world.chunks.addChunk({cx, cy, cz});
-                    //makeFlatTerrain(&chunk);
-                    makeNaturalTerrain(&chunk, heightMap);
-                }
+            for (int cy = 0; cy < 5; cy++) {
+                Chunk &chunk = m_world.chunks.addChunk({cx, cy, cz});
+                // makeFlatTerrain(&chunk);
+                makeNaturalTerrain(&chunk, heightMap);
+            }
         }
     }
     std::cout << "done\n";
