+++ src/server/world/terrain_generation.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <array>
-#include <common/world/world_constants.h>
 #include <common/world/coordinate.h>
+#include <common/world/world_constants.h>
 
 class Chunk;
 
@@ -10,7 +10,7 @@ void makeFlatTerrain(Chunk *chunk);
 void makeStepTerrain(Chunk *chunk);
 void makeRandomTerrain(Chunk *chunk);
 
-void makeNaturalTerrain(Chunk* chunk, const std::array<int, CHUNK_AREA>& heightMap);
-
+void makeNaturalTerrain(Chunk *chunk,
+                        const std::array<int, CHUNK_AREA> &heightMap);
 
 std::array<int, CHUNK_AREA> makeHeightMap(const ChunkPosition &location);
